Moves os_error_code out into include/flow
