+++ library/source/flow/utility.cpp
@@ -18,12 +18,12 @@
 #include <sys/stat.h> // for mkfifo
 #include <unistd.h> // for ::close
 
-#include "os_error_code.hpp"
 #include "wait_result.hpp"
 
 #include "flow/connection.hpp"
 #include "flow/descriptor_id.hpp"
 #include "flow/instance.hpp"
+#include "flow/os_error_code.hpp"
 #include "flow/system.hpp"
 #include "flow/utility.hpp"
 #include "flow/variant.hpp" // for <variant>, flow::variant, ostream support
